Handle map fields that are not sub-collections
Test data structure does not distinguish between a field that is a map and a field that represents a firestore sub-collection. Tests if the map has a length and if the first MapEntry also contains a Map. If it is a map of a map, then assume this is a sub-collection. Does not handle case in which the field is a map and that map contains a field that is a map.